requirements.txt
Django
djangorestframework
Pillow
torch  # or tensorflow
transformers
django-cors-headers
